Refactor Profile Page with Enhanced User Experience and Image Management
- Modularized Profile page into separate components for better code organization
- Added comprehensive error handling and loading states for profile image upload
- Implemented dynamic user data updates and stats rendering
- Enhanced ProfileImageModal with improved file validation and upload process
- Added uploading and error states to profile image actions
- Restructured profile page layout with more focused, reusable components
